requirements.txt
opencv-python-headless
cuda-python
datasets==2.16.1
hydra-core==1.3.2
#librosa==0.10.0
lightning==2.1.3
matplotlib==3.7.2
numpy==1.24.4
omegaconf==2.3.0
Pillow==10.0.0
pytest==7.4.0
scikit-image==0.21.0
scipy==1.11.1
torch~=2.2.0
torchmetrics~=1.3.1
torchvision~=0.17.0
torchaudio~=2.2.0
torch-audiomentations==0.11.0

rich~=13.7.0
transformers~=4.38.0
wandb~=0.16.0
librosa~=0.10.1
audiomentations~=0.32.0
soundfile~=0.12.1
pandas~=2.2.0
tqdm~=4.66.2
setuptools~=69.1.0
torchlibrosa~=0.1.0
safetensors~=0.4.2